tests are now passing after a medium refactor to the code structure (hexagon adapter for terraform)
