+++ lib/adapters/terraform/adapter.js
@@ -2,9 +2,12 @@
 // binary
 
 const fs = require('fs');
+const util = require('util');
 const writeFile = util.promisify(fs.writeFile);
-const { _exec } = require('child-process-promise');
+const { exec: _exec } = require('child-process-promise');
 const { isError } = require('lodash');
+const { spawn } = require('child_process');
+const path = require('path');
 
 async function exec(cmd, opts = {}, verbose = true) {
     if (verbose) {
@@ -24,17 +27,16 @@ async function exec(cmd, opts = {}, verbose = true) {
         console.log('stderr: ', stderr);
     }
 
-    return { _result, exitCode, stderr, stdout };
+    return { result, exitCode, stderr, stdout };
 }
 
 const adapter = {
-    async createSpinContextDirectory(contextId) {
-        const target = path.join(tmpBasePath, contextId);
+    async createSpinContextDirectory({ target }) {
         return exec(`mkdir -p ${target}`);
     },
     writeFile,
     async copyInfraBaseTemplate({ sourcePath, targetPath }) {
-        return exec(`cp -R ${sourcePath}/* ${targetPath} && rm -f eip.tf`, { cwd: sourcePath });
+        return exec(`cp -R ${sourcePath}/* ${targetPath} && rm -f ${path.join(targetPath,'eip.tf')}`, { cwd: sourcePath });
     },
     async copyInfraEipTemplate({ sourcePath, targetPath }) {
         return exec(`cp ${sourcePath}/eip.tf ${targetPath} && touch ${path.join(targetPath, '.eip')}`, { cwd: sourcePath });
@@ -111,28 +113,16 @@ const adapter = {
             });
         })
     },
-    apply({ targetPath }) {
+    apply({ targetPath, stdoutHandler }) {
         return new Promise((resolve, reject) => {
-            let outputs = [];
-            let outputsStarted = false;
-
             const tfApplySpawn = spawn('terraform', ['apply', '-var-file=terraform.tfvars', '-auto-approve'], {
                 cwd: targetPath,
             });
 
             tfApplySpawn.stdout.on('data', (data) => {
-                // Search for the public IP of the cluster manager node.
                 const dataAsString = data.toString();
-
-                if (dataAsString.indexOf('Outputs:') !== -1) {
-                    outputsStarted = true;
-                }
-
-                if (outputsStarted && dataAsString.indexOf(' = ') !== -1) {
-                    outputs = outputs.concat(this.parseOutputs(dataAsString));
-                }
-
                 console.log(dataAsString);
+                stdoutHandler(dataAsString);
             });
 
             tfApplySpawn.stderr.on('data', (data) => {
@@ -141,14 +131,14 @@ const adapter = {
 
             tfApplySpawn.on('close', (code) => {
                 if (code === 0) {
-                    resolve({ code, outputs });
+                    resolve({ code });
                     return;
                 }
                 reject({ code });
             });
         })
     },
-    checkFileExists({ targetPath }) {
+    checkFileExists(targetPath) {
         return fs.existsSync(targetPath);
     },
     detachElasticIPFromTerraformState({ targetPath }) {
