+++ lib/services/core/core.js
@@ -10,7 +10,7 @@ class CoreService {
 
     async createConstellation({ cloud = {}, keys = {}, }) {
         const eip = 'ip' in cloud;
-        const result = await this.terraformServiceInstancespinUp({
+        const result = await this.terraformServiceInstance.spinUp({
             cloud,
             keys,
         });
@@ -40,7 +40,7 @@ class CoreService {
     }
 
     async destroyConstellation({ spinContext }) {
-        const spinDownResult = await this.terraformServiceInstancespinDown({
+        const spinDownResult = await this.terraformServiceInstance.spinDown({
             spinContext,
         });
