+++ lib/services/terraform/terraform.js
@@ -15,9 +15,7 @@ class TerraformService {
     }
 
     async createSpinContextDirectory(contextId) {
-        const { childProcess: {
-            exitCode,
-        } } = await this.adapter.createSpinContextDirectory(contextId);
+        const { exitCode } = await this.adapter.createSpinContextDirectory(contextId);
 
         if (exitCode !== 0) {
             throw new Error("Couldn't create execution context directory for Terraform!");
@@ -174,13 +172,26 @@ class TerraformService {
 
     async apply({ spinContext }) {
         const targetPath = this.contextDirPath(spinContext);
+        let outputs = [];
+        let outputsStarted = false;
 
-        const result = await this.adapter.apply({ targetPath }).catch(err => err);
+        const stdoutHandler = (dataAsString) => {
+            // Search for the public IP of the cluster manager node.
+            if (dataAsString.indexOf('Outputs:') !== -1) {
+                outputsStarted = true;
+            }
+
+            if (outputsStarted && dataAsString.indexOf(' = ') !== -1) {
+                outputs = outputs.concat(this.parseOutputs(dataAsString));
+            }
+        };
+
+        const result = await this.adapter.apply({ targetPath, stdoutHandler }).catch(err => err);
         if (result.code !== 0) {
             return new Error('Could not perform Terraform apply phase!');
         }
 
-        return result;
+        return Object.assign({}, result, { outputs });
     }
 
     async destroy({ spinContext }) {
@@ -190,7 +201,7 @@ class TerraformService {
             console.log('Detaching Elastic IP from Terraform context...');
             // Looks like an Elastic IP was imported to this Terraform build
             // Let's detach it from Terraform so that we don't destroy it.
-            const detachResult = this.adapter.detachElasticIPFromTerraformState({ targetPath });
+            const detachResult = await this.adapter.detachElasticIPFromTerraformState({ targetPath });
             if (isError(detachResult)) {
                 console.error('Could not detach existing Elastic IP from Terraform state!');
                 return detachResult;
@@ -209,7 +220,8 @@ class TerraformService {
         const eip = 'ip' in cloud;
         const spinContext = uuid();
 
-        await this.createSpinContextDirectory(spinContext);
+        const target = path.join(tmpBasePath, spinContext);
+        await this.createSpinContextDirectory({ target });
         await this.writeTerraformVariablesFile({ spinContext, cloud, keys });
         await this.copyTerraformInfraTemplate({ cloud, spinContext });
         await this.copyStackScripts({ spinContext });
@@ -264,7 +276,7 @@ class TerraformService {
     }
 
     async spinDown({ spinContext }) {
-        const result = await this.adapter.destroy({ spinContext });
+        const result = await this.destroy({ spinContext });
         if (isError(result)) {
             return {
                 ok: false,
