+++ resources/terraform/aws/security-groups.tf
@@ -1,7 +1,8 @@
 /* Default security group */
 resource "aws_security_group" "swarm" {
-  name = "${var.application}-swarm-sg"
+  name        = "${var.application}-swarm-sg"
   description = "Default security group that allows inbound and outbound traffic from all instances in the VPC"
+  vpc_id      = "${ module.vpc.id }"
 
   ingress {
     from_port   = "0"
@@ -12,16 +13,16 @@ resource "aws_security_group" "swarm" {
   }
 
   ingress {
-    from_port = 22
-    to_port   = 22
-    protocol  = "tcp"
+    from_port   = 22
+    to_port     = 22
+    protocol    = "tcp"
     cidr_blocks = ["0.0.0.0/0"]
   }
 
   ingress {
-    from_port = 8080
-    to_port   = 8080
-    protocol  = "tcp"
+    from_port   = 8080
+    to_port     = 8080
+    protocol    = "tcp"
     cidr_blocks = ["0.0.0.0/0"]
   }
 
@@ -32,14 +33,15 @@ resource "aws_security_group" "swarm" {
     cidr_blocks = ["0.0.0.0/0"]
     self        = true
   }
+
   egress {
-    from_port = 22
-    to_port   = 22
-    protocol  = "tcp"
+    from_port   = 22
+    to_port     = 22
+    protocol    = "tcp"
     cidr_blocks = ["0.0.0.0/0"]
   }
-  
-  tags { 
-    Name = "swarm-example" 
+
+  tags {
+    Name = "${var.application}-swarm-sg"
   }
 }
